requirements.txt
pytest==8.3.1
python-dotenv==1.0.1
Flask==3.0.3
